requirements.txt
##### Protobuf for Runtime stuff #####  
protobuf

##### Google Sheets #####
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

##### Flask server #####
flask-socketio
gevent-websocket

##### Type checking #####
typeguard

##### Config #####
pyyaml